change ids to Long autodelete added

+++ app/src/main/java/ru/butterbean/easyrent/database/dao/GuestDao.kt
@@ -23,6 +23,6 @@ interface GuestDao {
     fun readAllGuests(): LiveData<List<GuestData>>
 
     @Query("SELECT * FROM $TABLE_GUESTS_NAME WHERE id = :id")
-    fun getById(id:Int): GuestData
+    fun getById(id:Long): GuestData
 
 }
