+++ app/src/main/java/ru/butterbean/easyrent/database/dao/ReserveDao.kt
@@ -23,12 +23,12 @@ interface ReserveDao {
     fun readAllReserves(): LiveData<List<ReserveData>>
 
     @Query("SELECT * FROM $TABLE_RESERVES_NAME WHERE roomId= :roomId ORDER BY id ASC")
-    fun getReservesByRoomId(roomId: Int): LiveData<List<ReserveData>>
+    fun getReservesByRoomId(roomId: Long): LiveData<List<ReserveData>>
 
     @Query("SELECT * FROM $TABLE_RESERVES_NAME WHERE date('now','start of day')==date(dateCheckOut,'start of day') & roomId= :roomId ORDER BY dateCheckIn ASC")
-    fun getActualReservesByRoomId(roomId: Int): LiveData<List<ReserveData>>
+    fun getActualReservesByRoomId(roomId: Long): LiveData<List<ReserveData>>
 
     @Query("SELECT * FROM $TABLE_RESERVES_NAME WHERE date(roomId,'start of day')==date(dateCheckOut,'start of day') & roomId= :roomId ORDER BY dateCheckIn ASC")
-    fun getEqualseservesByRoomId(roomId: Int): LiveData<List<ReserveData>>
+    fun getEqualseservesByRoomId(roomId: Long): LiveData<List<ReserveData>>
 
 }
